% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mcn_from_cel.R
\name{estimate_mcn_from_cel}
\alias{estimate_mcn_from_cel}
\title{Estimate MCN from CEL files.}
\usage{
estimate_mcn_from_cel(
  cel_list,
  output_dir,
  apt_lib_dir,
  gc5_file,
  autosomal_markers = hq_autosomal_markers,
  mitochondrial_markers = hq_mitochondrial_markers,
  apt_exec_dir = NULL,
  penncnv_exec_dir = NULL,
  pc_used = 0.01,
  keep_tempfile = TRUE,
  ignore_existing_tempfiles = FALSE,
  correlated_pheno = NULL,
  correlation_direction = "+"
)
}
\arguments{
\item{cel_list}{A character vector containing paths to all CEL files to be included in MCN estimation.}

\item{output_dir}{Path to the output directory.}

\item{apt_lib_dir}{Path to the library directory of CAS Array for APT. The directory should contain several XML for APT analysis and a CSV as array annotation.}

\item{gc5_file}{Path to the gzipped gc model file from PennCNV. The file could be found in gc_file/hg38.gc5Base.txt.gz in your local PennCNV installation or be downloaded at https://github.com/WGLab/PennCNV/blob/master/gc_file/hg38.gc5Base.txt.gz.}

\item{autosomal_markers}{(Optional) A character vector containing names of high-quality autosomal markers to be included in MCN estimation. Default is \code{hq_autosomal_markers} which contains a set of pre-defined high quality autosomal markers.}

\item{mitochondrial_markers}{(Optional) A character vector containing names of high-quality mitochondrial markers to be included in MCN estimation. Default is \code{hq_mitochondrial_markers} which contains a set of pre-defined high quality mitochondrial markers.}

\item{apt_exec_dir}{(Optional) Path to the directory containing APT executables. APT could be downloaded at https://www.thermofisher.com/cn/zh/home/life-science/microarray-analysis/microarray-analysis-partners-programs/affymetrix-developers-network/affymetrix-power-tools.html. Default is \code{NULL} which means the APT executables will be searched from system \code{PATH}.}

\item{penncnv_exec_dir}{(Optional) Path to the directory containing PennCNV executables. PennCNV could be downloaded at https://github.com/WGLab/PennCNV/releases. Default is \code{NULL} which means the PennCNV executables will be searched from system \code{PATH}.}

\item{pc_used}{(Optional) A number indicating principal components to be used for confounders correction. Either a fraction between 0 and 1 indicating proportion of post-QC sample size or an positive integer indicating count of PCs to be used are accepted. Default is \code{0.01}.}

\item{keep_tempfile}{(Optional) Whether to keep the temporary files of intermediate results for debug purpose or further analysis? Default is \code{TRUE}.}

\item{ignore_existing_tempfiles}{(Optional) If detected existing temporary files in the output directory, whether to use the intermediate results or delete them before the analysis? Use intermediate results will save time significantly, but incomplete intermediate results could cause issues. Default is \code{FALSE}.}

\item{correlated_pheno}{(Optional) A data frame with columns "Sample_Name" and "Phenotype". The phenotype should be associated with MCN for the oriantation of the first PC in PCA of mitochondrial markers. Default is \code{NULL} which means the MCN PGS will be used for PC oriantation. Note that the association between MCN PGS and MCN estimates may not strong enough in small populations (e.g. less than a few hundred of samples) so an associated phenotype is highly recommended.}

\item{correlation_direction}{(Optional) A character indicating the direction of the association between the provided phenotype and MCN. Accepts "\code{+}" or "\code{-}" indicating positive or negative correlation between the phenotype and MCN. Default is "\code{+}".}
}
\description{
Generate MCN estimates from CEL files producted by CAS Array.
}
