% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pc_adjustment.R
\name{plot_pc_adjustment}
\alias{plot_pc_adjustment}
\title{Plot MCN Performance Against Number of PCs Included.}
\usage{
plot_pc_adjustment(output_dir, max_pc_num = 0.01, correlated_pheno = NULL)
}
\arguments{
\item{output_dir}{Path to the output directory of \code{plot_pc_adjustment} function with "tempdir" directory in it.}

\item{max_pc_num}{(Optional) A number indicating the maximum principal components to be used for plotting. Either a fraction between 0 and 1 indicating proportion of the sample size or an positive integer indicating count of PCs are accepted. Default is \code{0.01}.}

\item{correlated_pheno}{(Optional) A data frame with columns "Sample_Name" and "Phenotype". The phenotype should be associated with MCN for checking the minimum number of PCs needed to capture the confounding factors. Default is \code{NULL} which means the MCN PGS will be used for PC checking. Note that the association between MCN PGS and MCN estimates may not strong enough in small populations (e.g. less than a few hundred of samples) so an associated phenotype is highly recommended.}
}
\value{
A plot showing correlation between estimated MCN and external phenotype (or MCN PGS if no phenotype is provided) against number of PCs used for adjustment.
}
\description{
Plot the correlation between estimated MCN and external phenotype (or MCN PGS if no phenotype is provided) against number of PCs used for adjustment.
}
